requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
attrs==24.1.0
beautifulsoup4==4.11.2
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.0
click==8.1.7
demjson3==3.0.6
discord==2.3.2
discord.py==2.4.0
docopt-ng==0.8.1
free_proxy==1.1.1
frozenlist==1.4.1
greenlet==3.0.3
gTTS==2.5.2
idna==2.10
lxml==5.3.0
multidict==6.0.5
mutagen==1.44.0
psycopg2==2.9.9
pycparser==2.22
pycryptodomex==3.20.0
PyNaCl==1.5.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
setuptools==47.1.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.32
typing_extensions==4.12.2
unicode-slugify==0.1.5
Unidecode==1.3.8
urllib3==1.26.19
websockets==12.0
yarl==1.9.4
youtube-dl==2021.12.17
yt-dlp==2024.8.6
